refine structures of backends
