+++ bitsandbytes/backends/__init__.py
@@ -1,4 +1,4 @@
-import typing
+from typing import Dict
 import torch
 
 from bitsandbytes.cextension import COMPILED_WITH_CUDA
